\name{ant_community}
\alias{ant_community}
\docType{data}
\title{Ant community example}
\description{
  A sample community matrix of ant species abundances with 3 sites and 6 species. Rows represent sites, columns represent species.
}
\usage{data(ant_community)}
\format{
  A data frame with 3 rows and 6 columns:
  \describe{
    \item{Atta sexdens}{abundance of \emph{Atta sexdens}}
    \item{Camponotus atriceps}{abundance of \emph{Camponotus atriceps}}
    \item{Pheidole megacephala}{abundance of \emph{Pheidole megacephala}}
    \item{Wasmannia auropunctata}{abundance of \emph{Wasmannia auropunctata}}
    \item{Solenopsis saevissima}{abundance of \emph{Solenopsis saevissima}}
    \item{Nylanderia fulva}{abundance of \emph{Nylanderia fulva}}
  }
}
\source{Simulated data for package examples.}
\examples{
data(ant_community)
head(ant_community)
}
\keyword{datasets}
