% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{delabie_db}
\alias{delabie_db}
\alias{endemic_db}
\alias{exotic_db}
\alias{generic_db}
\alias{rarity_db}
\alias{silva_db}
\alias{silvestre_db}
\title{Internal datasets used by AntClassify.}
\format{
Data frames

An object of class \code{data.frame} with 49 rows and 2 columns.

An object of class \code{data.frame} with 19 rows and 2 columns.

An object of class \code{data.frame} with 45 rows and 2 columns.

An object of class \code{data.frame} with 120 rows and 3 columns.

An object of class \code{data.frame} with 58 rows and 2 columns.

An object of class \code{data.frame} with 66 rows and 2 columns.
}
\usage{
delabie_db

endemic_db

exotic_db

generic_db

rarity_db

silva_db

silvestre_db
}
\description{
Internal datasets used by AntClassify.
}
\keyword{datasets}
\keyword{internal}
