% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antclassify.R
\name{antclassify}
\alias{antclassify}
\title{AntClassify Full Ecological Pipeline}
\usage{
antclassify(comm)
}
\arguments{
\item{comm}{Community matrix (samples x species).}
}
\value{
A list containing results from all analyses.
}
\description{
Runs all ecological classification functions of the package.
}
