% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_guild_ants.R
\name{assign_guild_ants}
\alias{assign_guild_ants}
\title{Assign Functional Guilds for Ants}
\usage{
assign_guild_ants(comm)
}
\arguments{
\item{comm}{A community matrix where species are columns and samples are rows.}
}
\description{
Classifies ants into functional guilds using four different ecological databases.
}
