% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_endemic_atlantic_ants.R
\name{check_endemic_atlantic_ants}
\alias{check_endemic_atlantic_ants}
\title{Identify Endemic Ant Species (Atlantic Forest)}
\usage{
check_endemic_atlantic_ants(comm)
}
\arguments{
\item{comm}{A community matrix where species are columns and samples are rows.}
}
\description{
Checks a community matrix for ant species endemic to the Brazilian Atlantic Forest based on Silva et al. (2025).
}
