% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_exotic_ants.R
\name{check_exotic_ants}
\alias{check_exotic_ants}
\title{Identify Exotic Ant Species (Brazil)}
\usage{
check_exotic_ants(comm)
}
\arguments{
\item{comm}{A community matrix where species are columns and samples are rows.}
}
\description{
Checks a community matrix for known exotic ant species in Brazil sourced from Vieira (2025).
}
