% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rarity_atlantic_ants.R
\name{check_rarity_atlantic_ants}
\alias{check_rarity_atlantic_ants}
\title{Identify Ant Rarity Forms (Atlantic Forest)}
\usage{
check_rarity_atlantic_ants(comm)
}
\arguments{
\item{comm}{A community matrix where species are columns and samples are rows.}
}
\description{
Checks a community matrix for ant rarity forms in the Brazilian Atlantic Forest based on Silva et al. (2024).
}
